сделал отдельный файл с обмена данными с бд
